% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.imptree}
\alias{predict.imptree}
\title{Classification with Imprecise Probabilities}
\usage{
\method{predict}{imptree}(object, data, dominance = c("strong", "max"),
  utility = 0.65, ...)
}
\arguments{
\item{object}{An object of class \code{imptree}. See details.}

\item{data}{Data.frame containing observations to be predicted.
If \code{NULL} the observations in the training set of \code{"object"}
are employed.}

\item{dominance}{Dominace criterion to be applied when predicting
classes. This may either be \code{"strong"} (default) or \code{"max"}.
See details.}

\item{utility}{Utility for the utility based accuracy measure for a 
vacuous prediction result (default: 0.65).}

\item{\dots}{Additional arguments for data. May be \code{"weights"},
\code{"subset"}, \code{"na.action"}, any further are discarded.}
}
\value{
A named list containing predicted classes, predicted 
probability distribution and accuracy evaluation
\itemize{
\item{probintlist}{: List of the Imprecise Probability Distribution of the
class variable. One matrix per observation.}
\item{classes}{: Predicted class(es) of the observations as boolean matrix}
\item{evaluation}{: Result of accuracy evaluation
\itemize{
\item{nObs}{: Number of observations}
\item{deter}{: Determinacy}
\item{nObsIndet}{: Number of observations with indeterminate prediction}
\item{indetSize}{: Average number of classes when predicting 
indeterminate (\code{NA} when no indeterminate observation)}
\item{acc_single}{: Single-set accuracy (\code{NA} when no determinate 
observation)}
\item{acc_set}{: Set-accuracy (\code{NA} when no indeterminate observation)}
\item{acc_disc}{: Discounted-accuracy}
\item{acc_util}{: Utility based (discounted) accuracy}
}}
}
}
\description{
Prediction of \code{imptree} objects
}
\details{
This function carries out the prediction of an imprecise tree. 
An existence check on the stored C++ object reference is carried out 
at first. If the reference is not valid the original call
for \code{"object"} is printed as error.

There are currently 2 different dominance criteria available.
\itemize{
\item{\code{max}}{: Maximum frequency criterion. Dominance is decided only
by the upper bound of the probability interval, ie. a state \eqn{C_i} is 
dominated if there exists any \eqn{j \neq i}{j != i} with 
\eqn{u(C_i) < u(C_j)}}.
\item{\code{strong}}{: Interval dominance criterion. For the IDM it
coincides with the strong dominance criterion. Here a state
\eqn{C_i} is dominated if there exists any \eqn{j \neq i}{j != i} 
with \eqn{u(C_i) < l(C_j)}}
}
}
\seealso{
\code{\link{imptree}}
}
\author{
Paul Fink \email{Paul.Fink@stat.uni-muenchen.de}
}
\keyword{tree}
