% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.imptree}
\alias{print.imptree}
\title{Classification with Imprecise Probabilities}
\usage{
\method{print}{imptree}(x, dominance = c("strong", "max"), ...)
}
\arguments{
\item{x}{Object of class \code{imptree}. See details.}

\item{dominance}{Dominace criterion to be applied when predicting
classes. See \code{\link{predict.imptree}} for details.}

\item{\dots}{Additional arguments; ignored at the moment}
}
\value{
Returns the calling object invisible.
}
\description{
Printing the \code{imptree} object to console
}
\details{
An existence check on stored Java Object References is carried out 
at first. If the Java Object References are not valid the original call
for \code{"object"} is printed as error.

For a more detailed summary of the tree \code{\link{summary.imptree}}.
}
\seealso{
\code{\link{imptree}}, \code{\link{summary.imptree}}
}
\author{
Paul Fink \email{Paul.Fink@stat.uni-muenchen.de}
}
\keyword{tree}
